requirements.txt
streamlit==1.35.0
transformers==4.41.2
torch
Pillow
timm
scipy
einops
numpy==1.26.4
easyocr
paddleocr